requirements/common.txt
Django
djangorestframework
djangorestframework-simplejwt
djangorestframework-simplejwt[crypto]
django-environ
# django-filter
Pillow
drf-nested-routers
django-cors-headers
# reportlab